research_agent/pyproject.toml
[project]
name = "research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "ramseywise", email = "ramsey.wise@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.88.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "structlog>=25.5.0",
]

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]

[tool.pytest.ini_options]
markers = [
    "integration: tests that hit real binaries or external services",
]
